Move ActionPage into Page directory.
