+++ resources/assets/components/Page/ActionPage/actionPage.scss
@@ -1,4 +1,4 @@
-@import '../../scss/next-toolbox.scss';
+@import '../../../scss/next-toolbox.scss';
 
 .action-step {
   width: 100%;
